Moved: Render from just.fp to maven2sbt.core
